Fix for error "IllegalArgumentException: Text must not be null or empty"
This error was caused intermittently when using kubectl to get the service account bearer token from a secret. We were encountering this issue: https://github.com/kubernetes/kubectl/issues/30 where the kubectl command would contain the error messages mentioned in the githus issue. This caused the secret token extraction from an expected json form to fail.
As a fix, I am extracting the token in the command itself. At the same time, I am avoiding the whole temporary file creation on the master node and the EF agent which was a security issue.